RDM-10413 - Rename to ccd-message-publisher
