+++ Jenkinsfile_CNP
@@ -4,7 +4,7 @@
 
 def type = "java"
 def product = "ccd"
-def component = "event-publisher"
+def component = "message-publisher"
 
 withPipeline(type, product, component) {
     disableLegacyDeployment()
