+++ build.gradle
@@ -1,13 +1,12 @@
 plugins {
   id 'application'
-  id 'checkstyle'
-  id 'pmd'
   id 'jacoco'
   id 'io.spring.dependency-management' version '1.0.10.RELEASE'
   id 'org.springframework.boot' version '2.3.4.RELEASE'
   id 'org.owasp.dependencycheck' version '6.0.3'
   id 'com.github.ben-manes.versions' version '0.36.0'
   id 'org.sonarqube' version '3.0'
+  id 'uk.gov.hmcts.java' version '0.11.0'
 }
 
 group = 'uk.gov.hmcts.reform'
@@ -84,21 +83,6 @@ task smoke(type: Test) {
   classpath = sourceSets.smokeTest.runtimeClasspath
 }
 
-checkstyle {
-  maxWarnings = 0
-  toolVersion = '8.29'
-  getConfigDirectory().set(new File(rootDir, 'config/checkstyle'))
-}
-
-pmd {
-  toolVersion = "6.18.0"
-  sourceSets = [sourceSets.main, sourceSets.test, sourceSets.functionalTest, sourceSets.integrationTest, sourceSets.smokeTest]
-  reportsDir = file("$project.buildDir/reports/pmd")
-  // https://github.com/pmd/pmd/issues/876
-  ruleSets = []
-  ruleSetFiles = files("config/pmd/ruleset.xml")
-}
-
 jacocoTestReport {
   executionData(test, integration)
   reports {
@@ -112,8 +96,8 @@ project.tasks['sonarqube'].dependsOn jacocoTestReport
 
 sonarqube {
   properties {
-    property "sonar.projectName", "Reform :: ccd-event-publisher"
-    property "sonar.projectKey", "uk.gov.hmcts.reform:ccd-event-publisher"
+    property "sonar.projectName", "Reform :: ccd-message-publisher"
+    property "sonar.projectKey", "uk.gov.hmcts.reform:ccd-message-publisher"
     property "sonar.coverage.jacoco.xmlReportPaths", "${jacocoTestReport.reports.xml.destination.path}"
   }
 }
@@ -214,7 +198,7 @@ mainClassName = 'uk.gov.hmcts.reform.uk.gov.hmcts.ccd.Application'
 
 bootJar {
   getArchiveFileName().set(provider {
-    'ccd-event-publisher.jar'
+    'ccd-message-publisher.jar'
   })
 
   manifest {
