+++ charts/ccd-message-publisher/templates/NOTES.txt
