+++ charts/ccd-message-publisher/values.yaml
@@ -1,0 +1,5 @@
+java:
+  applicationPort: 4455
+  image: 'hmctspublic.azurecr.io/ccd/message-publisher:latest'
+  ingressHost: ccd-message-publisher-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal
+  environment:
