+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '2.1'
 
 services:
-  ccd-event-publisher:
+  ccd-message-publisher:
     build:
       context: .
       args:
