+++ src/main/java/uk/gov/hmcts/ccd/controller/RootController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.uk.gov.hmcts.ccd.controllers;
+package uk.gov.hmcts.ccd.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,6 +23,6 @@ public class RootController {
      */
     @GetMapping("/")
     public ResponseEntity<String> welcome() {
-        return ok("Welcome to ccd-event-publisher");
+        return ok("Welcome to ccd-message-publisher");
     }
 }
